Move the index query context construction outside of OptRule
