+++ src/graph/optimizer/rule/IndexScanRule.h
@@ -12,11 +12,6 @@
 namespace nebula {
 namespace opt {
 
-using graph::QueryContext;
-using storage::cpp2::IndexColumnHint;
-using storage::cpp2::IndexQueryContext;
-using IndexItem = std::shared_ptr<meta::cpp2::IndexItem>;
-
 class OptContext;
 
 // Select the optimal index according to the preset filter inside [[IndexScan]]
@@ -33,131 +28,11 @@ class IndexScanRule final : public OptRule {
   std::string toString() const override;
 
  private:
-  struct ScanKind {
-    enum class Kind {
-      kUnknown = 0,
-      kMultipleScan,
-      kSingleScan,
-    };
-
-   private:
-    Kind kind_;
-
-   public:
-    ScanKind() {
-      kind_ = Kind::kUnknown;
-    }
-    void setKind(Kind k) {
-      kind_ = k;
-    }
-    Kind getKind() {
-      return kind_;
-    }
-    bool isSingleScan() {
-      return kind_ == Kind::kSingleScan;
-    }
-  };
-
-  // col_   : index column name
-  // relOP_ : Relational operator , for example c1 > 1 , the relOP_ == kRelGT
-  //                                            1 > c1 , the relOP_ == kRelLT
-  // value_ : Constant value. from ConstantExpression.
-  struct FilterItem {
-    std::string col_;
-    RelationalExpression::Kind relOP_;
-    Value value_;
-
-    FilterItem(const std::string& col, RelationalExpression::Kind relOP, const Value& value)
-        : col_(col), relOP_(relOP), value_(value) {}
-  };
-
-  // FilterItems used for optimal index fetch and index scan context optimize.
-  // for example : where c1 > 1 and c1 < 2 , the FilterItems should be :
-  //               {c1, kRelGT, 1} , {c1, kRelLT, 2}
-  struct FilterItems {
-    std::vector<FilterItem> items;
-    FilterItems() {}
-    explicit FilterItems(const std::vector<FilterItem>& i) {
-      items = i;
-    }
-    void addItem(const std::string& field, RelationalExpression::Kind kind, const Value& v) {
-      items.emplace_back(FilterItem(field, kind, v));
-    }
-  };
-
   static std::unique_ptr<OptRule> kInstance;
 
   IndexScanRule();
 
-  Status createIndexQueryCtx(std::vector<graph::IndexScan::IndexQueryContext>& iqctx,
-                             ScanKind kind,
-                             const FilterItems& items,
-                             graph::QueryContext* qctx,
-                             const OptGroupNode* groupNode) const;
-
-  Status createIndexQueryCtx(std::vector<graph::IndexScan::IndexQueryContext>& iqctx,
-                             graph::QueryContext* qctx,
-                             const OptGroupNode* groupNode) const;
-
-  Status createSingleIQC(std::vector<graph::IndexScan::IndexQueryContext>& iqctx,
-                         const FilterItems& items,
-                         graph::QueryContext* qctx,
-                         const OptGroupNode* groupNode) const;
-
-  Status createMultipleIQC(std::vector<graph::IndexScan::IndexQueryContext>& iqctx,
-                           const FilterItems& items,
-                           graph::QueryContext* qctx,
-                           const OptGroupNode* groupNode) const;
-
-  Status appendIQCtx(const IndexItem& index,
-                     const FilterItems& items,
-                     std::vector<graph::IndexScan::IndexQueryContext>& iqctx,
-                     const Expression* filter = nullptr) const;
-
-  Status appendIQCtx(const IndexItem& index,
-                     std::vector<graph::IndexScan::IndexQueryContext>& iqctx) const;
-
-  Status appendColHint(std::vector<IndexColumnHint>& hitns,
-                       const FilterItems& items,
-                       const meta::cpp2::ColumnDef& col) const;
-
-  size_t hintCount(const FilterItems& items) const;
-
-  bool isEdge(const OptGroupNode* groupNode) const;
-
-  int32_t schemaId(const OptGroupNode* groupNode) const;
-
-  GraphSpaceID spaceId(const OptGroupNode* groupNode) const;
-
   Expression* filterExpr(const OptGroupNode* groupNode) const;
-
-  Status analyzeExpression(
-      Expression* expr, FilterItems* items, ScanKind* kind, bool isEdge, QueryContext* qctx) const;
-
-  template <typename E,
-            typename = std::enable_if_t<std::is_same<E, EdgePropertyExpression>::value ||
-                                        std::is_same<E, LabelTagPropertyExpression>::value ||
-                                        std::is_same<E, TagPropertyExpression>::value>>
-  Status addFilterItem(RelationalExpression* expr, FilterItems* items, QueryContext* qctx) const;
-
-  IndexItem findOptimalIndex(graph::QueryContext* qctx,
-                             const OptGroupNode* groupNode,
-                             const FilterItems& items) const;
-
-  IndexItem findLightestIndex(graph::QueryContext* qctx, const OptGroupNode* groupNode) const;
-
-  std::vector<IndexItem> allIndexesBySchema(graph::QueryContext* qctx,
-                                            const OptGroupNode* groupNode) const;
-
-  std::vector<IndexItem> findValidIndex(graph::QueryContext* qctx,
-                                        const OptGroupNode* groupNode,
-                                        const FilterItems& items) const;
-
-  std::vector<IndexItem> findIndexForEqualScan(const std::vector<IndexItem>& indexes,
-                                               const FilterItems& items) const;
-
-  std::vector<IndexItem> findIndexForRangeScan(const std::vector<IndexItem>& indexes,
-                                               const FilterItems& items) const;
 };
 
 }  // namespace opt
