+++ src/graph/optimizer/test/IndexScanRuleTest.cpp
@@ -10,25 +10,26 @@
 
 using nebula::cpp2::PropertyType;
 using nebula::graph::OptimizerUtils;
+using nebula::storage::cpp2::IndexQueryContext;
+
+using FilterItem = nebula::graph::OptimizerUtils::FilterItem;
 
 namespace nebula {
 namespace opt {
 
 TEST(IndexScanRuleTest, IQCtxTest) {
-  auto* inst = std::move(IndexScanRule::kInstance).get();
-  auto* instance = static_cast<IndexScanRule*>(inst);
   auto objPoolPtr = std::make_unique<ObjectPool>();
   auto* pool = objPoolPtr.get();
   {
-    IndexItem index = std::make_unique<meta::cpp2::IndexItem>();
-    IndexScanRule::FilterItems items;
+    auto index = std::make_shared<meta::cpp2::IndexItem>();
+    std::vector<FilterItem> items;
     std::vector<IndexQueryContext> iqctx;
-    auto ret = instance->appendIQCtx(index, items, iqctx);
+    auto ret = OptimizerUtils::appendIQCtx(index, items, iqctx);
     ASSERT_TRUE(ret.ok());
   }
   {
-    IndexItem index = std::make_unique<meta::cpp2::IndexItem>();
-    IndexScanRule::FilterItems items;
+    auto index = std::make_shared<meta::cpp2::IndexItem>();
+    std::vector<FilterItem> items;
     std::vector<IndexQueryContext> iqctx;
     // setup index
     {
@@ -44,11 +45,11 @@ TEST(IndexScanRuleTest, IQCtxTest) {
     }
     // setup FilterItems col0 < 1 and col0 <= 6
     {
-      items.items.clear();
-      items.addItem("col0", RelationalExpression::Kind::kRelLT, Value(1L));
-      items.addItem("col0", RelationalExpression::Kind::kRelLE, Value(6L));
+      items.clear();
+      items.emplace_back("col0", RelationalExpression::Kind::kRelLT, Value(1L));
+      items.emplace_back("col0", RelationalExpression::Kind::kRelLE, Value(6L));
 
-      auto ret = instance->appendIQCtx(index, items, iqctx);
+      auto ret = OptimizerUtils::appendIQCtx(index, items, iqctx);
       ASSERT_TRUE(ret.ok());
 
       ASSERT_EQ(1, iqctx.size());
@@ -66,16 +67,16 @@ TEST(IndexScanRuleTest, IQCtxTest) {
     // setup FilterItems col0 > 1 and col1 <= 2 and col1 > -1 and col2 > 3
     //                   and col3 < 4 and col4 == 4
     {
-      items.items.clear();
+      items.clear();
       iqctx.clear();
-      items.addItem("col0", RelationalExpression::Kind::kRelGT, Value(1L));
-      items.addItem("col1", RelationalExpression::Kind::kRelLE, Value(2L));
-      items.addItem("col1", RelationalExpression::Kind::kRelGT, Value(-1L));
-      items.addItem("col2", RelationalExpression::Kind::kRelGT, Value(3L));
-      items.addItem("col3", RelationalExpression::Kind::kRelLT, Value(4L));
-      items.addItem("col4", RelationalExpression::Kind::kRelEQ, Value(4L));
+      items.emplace_back("col0", RelationalExpression::Kind::kRelGT, Value(1L));
+      items.emplace_back("col1", RelationalExpression::Kind::kRelLE, Value(2L));
+      items.emplace_back("col1", RelationalExpression::Kind::kRelGT, Value(-1L));
+      items.emplace_back("col2", RelationalExpression::Kind::kRelGT, Value(3L));
+      items.emplace_back("col3", RelationalExpression::Kind::kRelLT, Value(4L));
+      items.emplace_back("col4", RelationalExpression::Kind::kRelEQ, Value(4L));
 
-      auto ret = instance->appendIQCtx(index, items, iqctx);
+      auto ret = OptimizerUtils::appendIQCtx(index, items, iqctx);
       ASSERT_TRUE(ret.ok());
 
       ASSERT_EQ(1, iqctx.size());
@@ -98,15 +99,15 @@ TEST(IndexScanRuleTest, IQCtxTest) {
     // only expect col0 and col1 exits in column hints.
     // col2 and col3 should be filter in storage layer.
     {
-      items.items.clear();
+      items.clear();
       iqctx.clear();
-      items.addItem("col0", RelationalExpression::Kind::kRelEQ, Value(1L));
-      items.addItem("col1", RelationalExpression::Kind::kRelLE, Value(2L));
-      items.addItem("col1", RelationalExpression::Kind::kRelGT, Value(-1L));
-      items.addItem("col2", RelationalExpression::Kind::kRelNE, Value(3L));
-      items.addItem("col3", RelationalExpression::Kind::kRelLT, Value(4L));
+      items.emplace_back("col0", RelationalExpression::Kind::kRelEQ, Value(1L));
+      items.emplace_back("col1", RelationalExpression::Kind::kRelLE, Value(2L));
+      items.emplace_back("col1", RelationalExpression::Kind::kRelGT, Value(-1L));
+      items.emplace_back("col2", RelationalExpression::Kind::kRelNE, Value(3L));
+      items.emplace_back("col3", RelationalExpression::Kind::kRelLT, Value(4L));
 
-      auto ret = instance->appendIQCtx(index, items, iqctx);
+      auto ret = OptimizerUtils::appendIQCtx(index, items, iqctx);
       ASSERT_TRUE(ret.ok());
 
       ASSERT_EQ(1, iqctx.size());
@@ -135,17 +136,17 @@ TEST(IndexScanRuleTest, IQCtxTest) {
     // only expect col0, col1, col2 and col3 exits in column hints.
     // col4 should be filter in storage layer.
     {
-      items.items.clear();
+      items.clear();
       iqctx.clear();
-      items.addItem("col0", RelationalExpression::Kind::kRelEQ, Value(1L));
-      items.addItem("col1", RelationalExpression::Kind::kRelEQ, Value(2L));
-      items.addItem("col2", RelationalExpression::Kind::kRelEQ, Value(-1L));
-      items.addItem("col3", RelationalExpression::Kind::kRelGT, Value(3L));
-      items.addItem("col4", RelationalExpression::Kind::kRelLT, Value(4L));
+      items.emplace_back("col0", RelationalExpression::Kind::kRelEQ, Value(1L));
+      items.emplace_back("col1", RelationalExpression::Kind::kRelEQ, Value(2L));
+      items.emplace_back("col2", RelationalExpression::Kind::kRelEQ, Value(-1L));
+      items.emplace_back("col3", RelationalExpression::Kind::kRelGT, Value(3L));
+      items.emplace_back("col4", RelationalExpression::Kind::kRelLT, Value(4L));
 
       auto* expr = RelationalExpression::makeLT(
           pool, ConstantExpression::make(pool, "col4"), ConstantExpression::make(pool, 4));
-      auto ret = instance->appendIQCtx(index, items, iqctx, expr);
+      auto ret = OptimizerUtils::appendIQCtx(index, items, iqctx, expr);
       ASSERT_TRUE(ret.ok());
 
       ASSERT_EQ(1, iqctx.size());
@@ -183,9 +184,6 @@ TEST(IndexScanRuleTest, IQCtxTest) {
 }
 
 TEST(IndexScanRuleTest, BoundValueRangeTest) {
-  auto* inst = std::move(IndexScanRule::kInstance).get();
-  auto* instance = static_cast<IndexScanRule*>(inst);
-
   {
     meta::cpp2::ColumnDef col;
     col.name_ref() = "col_int";
@@ -193,9 +191,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_int < 2
-      IndexScanRule::FilterItems items;
-      items.addItem("col_int", RelationalExpression::Kind::kRelLT, Value(2L));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_int", RelationalExpression::Kind::kRelLT, Value(2L));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -208,9 +206,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_int <= 2
-      IndexScanRule::FilterItems items;
-      items.addItem("col_int", RelationalExpression::Kind::kRelLE, Value(2L));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_int", RelationalExpression::Kind::kRelLE, Value(2L));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -223,9 +221,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_int > 2
-      IndexScanRule::FilterItems items;
-      items.addItem("col_int", RelationalExpression::Kind::kRelGT, Value(2L));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_int", RelationalExpression::Kind::kRelGT, Value(2L));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -238,9 +236,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_int >= 2
-      IndexScanRule::FilterItems items;
-      items.addItem("col_int", RelationalExpression::Kind::kRelGE, Value(2L));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_int", RelationalExpression::Kind::kRelGE, Value(2L));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -253,10 +251,10 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_int > 2 and col_int < 5
-      IndexScanRule::FilterItems items;
-      items.addItem("col_int", RelationalExpression::Kind::kRelGT, Value(2L));
-      items.addItem("col_int", RelationalExpression::Kind::kRelLT, Value(5L));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_int", RelationalExpression::Kind::kRelGT, Value(2L));
+      items.emplace_back("col_int", RelationalExpression::Kind::kRelLT, Value(5L));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -270,10 +268,10 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_int >= 2 and col_int <= 5
-      IndexScanRule::FilterItems items;
-      items.addItem("col_int", RelationalExpression::Kind::kRelGE, Value(2L));
-      items.addItem("col_int", RelationalExpression::Kind::kRelLE, Value(5L));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_int", RelationalExpression::Kind::kRelGE, Value(2L));
+      items.emplace_back("col_int", RelationalExpression::Kind::kRelLE, Value(5L));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -292,17 +290,17 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_bool < true
-      IndexScanRule::FilterItems items;
-      items.addItem("col_bool", RelationalExpression::Kind::kRelLT, Value(true));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_bool", RelationalExpression::Kind::kRelLT, Value(true));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_FALSE(status.ok());
     }
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_bool >= false
-      IndexScanRule::FilterItems items;
-      items.addItem("col_bool", RelationalExpression::Kind::kRelGE, Value(false));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_bool", RelationalExpression::Kind::kRelGE, Value(false));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_FALSE(status.ok());
     }
   }
@@ -313,9 +311,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_double < 1.0
-      IndexScanRule::FilterItems items;
-      items.addItem("col_double", RelationalExpression::Kind::kRelLT, Value(1.0));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_double", RelationalExpression::Kind::kRelLT, Value(1.0));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -328,9 +326,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_double <= 3.0
-      IndexScanRule::FilterItems items;
-      items.addItem("col_double", RelationalExpression::Kind::kRelLE, Value(3.0));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_double", RelationalExpression::Kind::kRelLE, Value(3.0));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -343,9 +341,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_double >= 1.0
-      IndexScanRule::FilterItems items;
-      items.addItem("col_double", RelationalExpression::Kind::kRelGE, Value(1.0));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_double", RelationalExpression::Kind::kRelGE, Value(1.0));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -358,10 +356,10 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // 1.0 < col_double <= 5.0
-      IndexScanRule::FilterItems items;
-      items.addItem("col_double", RelationalExpression::Kind::kRelGT, Value(1.0));
-      items.addItem("col_double", RelationalExpression::Kind::kRelLE, Value(5.0));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_double", RelationalExpression::Kind::kRelGT, Value(1.0));
+      items.emplace_back("col_double", RelationalExpression::Kind::kRelLE, Value(5.0));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -382,9 +380,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_str < "ccc"
-      IndexScanRule::FilterItems items;
-      items.addItem("col_str", RelationalExpression::Kind::kRelLT, Value("ccc"));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_str", RelationalExpression::Kind::kRelLT, Value("ccc"));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -397,9 +395,9 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // col_str > "ccc"
-      IndexScanRule::FilterItems items;
-      items.addItem("col_str", RelationalExpression::Kind::kRelGT, Value("ccc"));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_str", RelationalExpression::Kind::kRelGT, Value("ccc"));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -412,10 +410,10 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // "aaa" < col_str < "ccc"
-      IndexScanRule::FilterItems items;
-      items.addItem("col_str", RelationalExpression::Kind::kRelGT, Value("aaa"));
-      items.addItem("col_str", RelationalExpression::Kind::kRelLT, Value("ccc"));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_str", RelationalExpression::Kind::kRelGT, Value("aaa"));
+      items.emplace_back("col_str", RelationalExpression::Kind::kRelLT, Value("ccc"));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -429,10 +427,10 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // "aaa" <= col_str <= "ccc"
-      IndexScanRule::FilterItems items;
-      items.addItem("col_str", RelationalExpression::Kind::kRelGE, Value("aaa"));
-      items.addItem("col_str", RelationalExpression::Kind::kRelLE, Value("ccc"));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_str", RelationalExpression::Kind::kRelGE, Value("aaa"));
+      items.emplace_back("col_str", RelationalExpression::Kind::kRelLE, Value("ccc"));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
@@ -446,10 +444,10 @@ TEST(IndexScanRuleTest, BoundValueRangeTest) {
     {
       std::vector<storage::cpp2::IndexColumnHint> hints;
       // "aaa" <= col_str < "ccc"
-      IndexScanRule::FilterItems items;
-      items.addItem("col_str", RelationalExpression::Kind::kRelGE, Value("aaa"));
-      items.addItem("col_str", RelationalExpression::Kind::kRelLT, Value("ccc"));
-      auto status = instance->appendColHint(hints, items, col);
+      std::vector<FilterItem> items;
+      items.emplace_back("col_str", RelationalExpression::Kind::kRelGE, Value("aaa"));
+      items.emplace_back("col_str", RelationalExpression::Kind::kRelLT, Value("ccc"));
+      auto status = OptimizerUtils::appendColHint(hints, items, col);
       EXPECT_TRUE(status.ok());
       EXPECT_EQ(1, hints.size());
       const auto& hint = hints[0];
