+++ src/graph/planner/match/MatchSolver.cpp
@@ -213,91 +213,6 @@ Expression* MatchSolver::makeIndexFilter(const std::string& label,
   return root;
 }
 
-// Expression* MatchSolver::rewriteTagIndexFilter(const std::string& label,
-//                                                const std::string& alias,
-//                                                Expression* filter,
-//                                                QueryContext* qctx,
-//                                                std::string* varName) {
-//   auto kind = filter->kind();
-//   if (kind != Expression::Kind::kRelEQ && kind != Expression::Kind::kLogicalAnd) {
-//     return nullptr;
-//   }
-
-//   std::vector<const Expression*> ands;
-//   if (kind == Expression::Kind::kLogicalAnd) {
-//     auto* logic = static_cast<LogicalExpression*>(filter);
-//     ExpressionUtils::pullAnds(logic);
-//     for (auto& operand : logic->operands()) {
-//       if (operand->kind() == Expression::Kind::kRelEQ ||
-//           operand->kind() == Expression::Kind::kRelIn) {
-//         ands.emplace_back(operand);
-//       }
-//     }
-//   } else {
-//     ands.emplace_back(filter);
-//   }
-
-//   std::string propName;
-//   auto checkExpr = [varName, &alias, &label, &propName](const Expression* expr,
-//                                                         const Expression* varExpr) -> bool {
-//     if (expr->kind() != Expression::Kind::kLabelTagProperty ||
-//         varExpr->kind() != Expression::Kind::kLabel) {
-//       return false;
-//     }
-
-//     auto laExpr = static_cast<const LabelTagPropertyExpression*>(expr);
-//     if (laExpr->sym() != label) return false;
-
-//     auto propExpr = dynamic_cast<const PropertyExpression*>(laExpr->label());
-//     if (!propExpr || propExpr->prop() != alias) return false;
-
-//     auto labelExpr = static_cast<const LabelExpression*>(varExpr);
-//     *varName = labelExpr->name();
-//     propName = laExpr->prop();
-
-//     return true;
-//   };
-
-//   auto* pool = qctx->objPool();
-//   std::vector<Expression*> relationals;
-//   for (auto* item : ands) {
-//     if (item->kind() != Expression::Kind::kRelEQ) {
-//       continue;
-//     }
-
-//     auto* binary = static_cast<const RelationalExpression*>(item);
-//     auto* left = binary->left();
-//     auto* right = binary->right();
-//     if (left->kind() == Expression::Kind::kLabelTagProperty) {
-//       if (!checkExpr(left, right)) continue;
-//     } else if (right->kind() == Expression::Kind::kLabelTagProperty) {
-//       if (!checkExpr(right, left)) continue;
-//     } else {
-//       continue;
-//     }
-
-//     auto tpExpr = TagPropertyExpression::make(pool, label, propName);
-//     if (right->kind() == Expression::Kind::kConstant) {
-//       auto* rel = RelationalExpression::makeKind(pool, item->kind(), tpExpr, newConstant);
-//       relationals.emplace_back(rel);
-//     } else {
-//       auto* rel = RelationalExpression::makeKind(pool, item->kind(), newConstant, tpExpr);
-//       relationals.emplace_back(rel);
-//     }
-//   }
-
-//   if (relationals.empty()) {
-//     return nullptr;
-//   }
-
-//   auto* root = relationals[0];
-//   for (auto i = 1u; i < relationals.size(); i++) {
-//     root = LogicalExpression::makeAnd(qctx->objPool(), root, relationals[i]);
-//   }
-
-//   return root;
-// }
-
 PlanNode* MatchSolver::filtPathHasSameEdge(PlanNode* input,
                                            const std::string& column,
                                            QueryContext* qctx) {
